requirements.txt
streamlit
pandas
fpdf2
qrcode
pillow
openpyxl